+++ src/apps/ScopeChangeRequest/api/PCS/Batch/index.ts
@@ -1,0 +1,2 @@
+export * from './batchCommPkg';
+export * from './batchTag';
